Fix synchronization bug in VisualContentMAnager

+++ app/src/main/java/com/projecttango/rajawali/ContentPlane.java
@@ -191,21 +191,21 @@ public class ContentPlane extends Object3D {
                 int lr = ur + 1;
 
                 if (mUpAxis == Vector3.Axis.X || mUpAxis == Vector3.Axis.Z) {
-                    indices[indexCount++] = (int) ur;
-                    indices[indexCount++] = (int) lr;
-                    indices[indexCount++] = (int) ul;
+                    indices[indexCount++] = ur;
+                    indices[indexCount++] = lr;
+                    indices[indexCount++] = ul;
 
-                    indices[indexCount++] = (int) lr;
-                    indices[indexCount++] = (int) ll;
-                    indices[indexCount++] = (int) ul;
+                    indices[indexCount++] = lr;
+                    indices[indexCount++] = ll;
+                    indices[indexCount++] = ul;
                 } else {
-                    indices[indexCount++] = (int) ur;
-                    indices[indexCount++] = (int) ul;
-                    indices[indexCount++] = (int) lr;
+                    indices[indexCount++] = ur;
+                    indices[indexCount++] = ul;
+                    indices[indexCount++] = lr;
 
-                    indices[indexCount++] = (int) lr;
-                    indices[indexCount++] = (int) ul;
-                    indices[indexCount++] = (int) ll;
+                    indices[indexCount++] = lr;
+                    indices[indexCount++] = ul;
+                    indices[indexCount++] = ll;
                 }
             }
         }
